Extract common things to `cardano-ledger-core`.
There were a couple of `WellFormed` constraints dropped here, since they
would entail pulling in a lot of Shelley-era dependencies. It may be
possible to refactor to regain these, but we can do that in a follow-up
if needed.
