+++ hie.yaml
@@ -33,6 +33,9 @@ cradle:
     - path: "semantics/small-steps-test/test"
       component: "small-steps-test:test:examples"
 
+    - path: "cardano-ledger-core/src"
+      component: "lib:cardano-ledger-core"
+
     - path: "shelley/chain-and-ledger/dependencies/non-integer/src"
       component: "lib:shelley-spec-non-integral"
