--- shelley/chain-and-ledger/executable-spec/src/Cardano/Ledger/Torsor.hs
@@ -1,15 +1,0 @@
-{-# LANGUAGE DeriveGeneric #-}
-{-# LANGUAGE DerivingVia #-}
-{-# LANGUAGE FlexibleInstances #-}
-{-# LANGUAGE GeneralizedNewtypeDeriving #-}
-{-# LANGUAGE OverloadedStrings #-}
-{-# LANGUAGE TypeFamilies #-}
-
-module Cardano.Ledger.Torsor where
-
-import Data.Kind (Type)
-
-class Torsor a where
-  type Delta a :: Type
-  addDelta :: a -> Delta a -> a
-  toDelta :: a -> Delta a
